+++ src/core/prelude/number/format.ts
@@ -80,7 +80,7 @@ extend(Number.prototype, 'format', function (
 		return this.toLocaleString(locale);
 	}
 
-	if (Object.isObject(patternOrOpts)) {
+	if (Object.isPlainObject(patternOrOpts)) {
 		return this.toLocaleString(locale, patternOrOpts);
 	}